+++ cypress/integration/granules_spec.js
@@ -419,7 +419,7 @@ describe('Dashboard Granules Page', () => {
     it('Should reingest a granule from granule detail page.', () => {
       const granuleId = 'MOD09GQ.A9344328.K9yI3O.006.4625818663028';
       cy.intercept(
-        { method: 'PUT', url: /\/granules\/.*/ },
+        { method: 'PATCH', url: /\/granules\/.*/ },
         { body: { message: 'ingested' }, statusCode: 200 },
       );
       cy.intercept('GET', `/granules/${granuleId}*`).as('getGranule');
@@ -441,7 +441,7 @@ describe('Dashboard Granules Page', () => {
     it('Should reingest a granule and redirect to the granule detail page.', () => {
       const granuleId = 'MOD09GQ.A0142558.ee5lpE.006.5112577830916';
       cy.intercept(
-        { method: 'PUT', url: /\/granules\/.*/ },
+        { method: 'PATCH', url: /\/granules\/.*/ },
         { body: { message: 'ingested' }, statusCode: 200 },
       );
       cy.intercept('GET', `/granules/${granuleId}*`).as('getGranule');
@@ -466,7 +466,7 @@ describe('Dashboard Granules Page', () => {
         'MOD09GQ.A9344328.K9yI3O.006.4625818663028'
       ];
       cy.intercept(
-        { method: 'PUT', url: /\/granules\/.*/ },
+        { method: 'PATCH', url: /\/granules\/.*/ },
         { body: { message: 'ingested' }, statusCode: 200 },
       );
       cy.visit('/granules');
@@ -486,7 +486,7 @@ describe('Dashboard Granules Page', () => {
 
     it('Should reingest multiple granules selected from multiple pages.', () => {
       cy.intercept(
-        { method: 'PUT', url: /\/granules\/.*/ },
+        { method: 'PATCH', url: /\/granules\/.*/ },
         { body: { message: 'ingested' }, statusCode: 200 },
       );
       cy.visit('/granules?limit=2');
@@ -519,7 +519,7 @@ describe('Dashboard Granules Page', () => {
         'MOD09GQ.A9344328.K9yI3O.006.4625818663028'
       ];
       cy.intercept(
-        { method: 'PUT', url: /\/granules\/.*/ },
+        { method: 'PATCH', url: /\/granules\/.*/ },
         { body: { message: 'Oopsie' }, statusCode: 500 }
       );
       cy.visit('/granules');
@@ -759,7 +759,7 @@ describe('Dashboard Granules Page', () => {
     });
 
     it('Should handle a successful API response from the Remove and Delete granule requests', () => {
-      cy.intercept({ method: 'PUT', url: /\/granules\/.*/ }, (req) => {
+      cy.intercept({ method: 'PATCH', url: /\/granules\/.*/ }, (req) => {
         expect(req.body).to.have.property('action', 'removeFromCmr');
         req.reply({ body: { message: 'success' }, statusCode: 200 });
       });